Allow path after a shortcut: go/short/foobar
This allows adding additional path to a shortened URL. For example, if
go/short redirects to http://example.com, go/short/foobar will redirect
to http://example.com/foobar.

An example usage is to shorten Google Docs headings. go/mydoc can
redirect to a document. go/mydoc/#heading=h.asdfgh123456 can then
redirect to a specific heading in the doc.

This feature is supported in the original golinks at Google.

Next step: Create a chrome extension that links to the go/ server and
shortens a given link based on the shortcuts already registered